backend/requirements.txt
ultralytics
opencv-python
fastapi
uvicorn
paho-mqtt
requests
python-dotenv
uvloop; sys_platform != 'win32'
python-multipart
websockets
